More fixes

+++ src/lib/storage/buffer/types.hpp
@@ -13,7 +13,7 @@
 #include <numa.h>
 #endif
 
-STRONG_TYPEDEF(int8_t, NumaMemoryNode);
+STRONG_TYPEDEF(int64_t, NumaMemoryNode);
 
 namespace hyrise {
